Add files via upload
In this project, I conducted a detailed cleaning and analysis of the financial consumer complaints filed against Cumulus Financial.
The analysis aimed to uncover trends in customer complaints, identify the most frequent issues reported, and assess the performance of Cumulus Financials' response to these complaints. By visualizing the data in Excel through pivot tables and charts, I was able to highlight the distribution of complaints across various products and geographic regions, as well as the speed and effectiveness of the company's resolution processes.